privacy/text_video/requirements.txt
accelerate==1.2.1
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
attrs==24.3.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
blinker==1.9.0
boto3==1.35.82
botocore==1.35.82
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colossalai==0.4.6
contexttimer==0.3.3
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
datasets==3.2.0
decorator==5.1.1
decord==0.6.0
Deprecated==1.2.15
diffusers @ git+https://github.com/huggingface/diffusers@ec1c7a793f9cdcb924d302f121348d9bb5256597
dill==0.3.8
distlib==0.3.9
distro==1.9.0
DistVAE==0.0.0b4
einops==0.8.0
fabric==3.2.2
fastapi==0.115.6
filelock==3.16.1
flash-attn==2.7.2.post1
Flask==3.1.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
galore-torch==1.0
git+https://github.com/openai/CLIP.git
google==3.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
identify==2.6.3
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
iniconfig==2.0.0
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lightning==2.4.0
lightning-utilities==0.11.9
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mmengine==0.10.5
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.3
nodeenv==1.9.1
numpy==2.2.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open-clip-torch==2.24.0
openai==1.57.4
opencv-python==4.10.0.84
optimum-quanto==0.2.6
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
peft==0.14.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
plumbum==1.9.0
pre_commit==4.0.1
proglog==0.1.10
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
pyarrow==18.1.0
pyav==12.3.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rotary-embedding-torch==0.8.6
rpds-py==0.22.3
rpyc==6.0.0
s3transfer==0.10.4
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
termcolor==2.5.0
timm==1.0.12
tokenizers==0.21.0
torch==2.4.0
torchmetrics==1.6.0
torchvision==0.19.0
tqdm==4.67.1
transformers==4.47.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.29.0
virtualenv==20.28.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
xformers==0.0.27.post2
xfuser==0.4.0
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
yunchang==0.5.0
zipp==3.21.0
